+++ frontend/package.json
@@ -14,7 +14,8 @@
     "aws-amplify": "^6.10.2",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
-    "react-helmet-async": "^2.0.5"
+    "react-helmet-async": "^2.0.5",
+    "react-router-dom": "^7.0.2"
   },
   "devDependencies": {
     "@eslint/eslintrc": "^3.2.0",
