+++ frontend/src/components/TicketTrack.tsx
@@ -1,26 +1,29 @@
-import React from 'react';
+import { Routes, Route } from 'react-router-dom';
 import { Helmet } from 'react-helmet-async';
 import Header from './header';
 import SidePanel from './main/sidePanel';
-import TicketsPage from './main/TicketsPage';
-import './TicketTrack.css';
+import TicketsPage from './main/ticketsPage';
+import './index.css';
 
+/**
+ * Represents the main component for the application.
+ */
 const TicketTrack = () => (
-  <>
-    <div className='App'>
-      <Helmet>
-        <title>Ticket Track</title>
-        <link rel='manifest' href='/manifest.json' />
-      </Helmet>
-      <Header />
-      <div className='main-layout'>
-        <SidePanel />
-        <div className='main-content'>
-          <TicketsPage />
-        </div>
+  <div className='app-container'>
+    <Helmet>
+      <title>Ticket Track</title>
+      <link rel='manifest' href='/manifest.json' />
+    </Helmet>
+    <Header />
+    <div className='main-layout'>
+      <SidePanel />
+      <div className='main-content'>
+        <Routes>
+          <Route path='/' element={<TicketsPage />} />
+        </Routes>
       </div>
     </div>
-  </>
+  </div>
 );
 
 export default TicketTrack;
