+++ frontend/src/components/header/index.tsx
@@ -1,5 +1,8 @@
 import './index.css';
 
+/**
+ * Represents the header component for the application, containing the logo and title.
+ */
 const Header = () => (
   <header className='header'>
     <img src={'/logo192.png'} alt='Logo' className='logo' />
