+++ frontend/src/components/index.css
@@ -1,4 +1,4 @@
-.App {
+.app-container {
   text-align: center;
   background-color: rgb(36, 36, 36);
   min-height: 100vh;
