+++ frontend/src/components/main/TicketsPage/index.css
@@ -1,9 +1,9 @@
-.search-info {
+.event-info-container {
   display: flex;
   flex-direction: row;
 }
 
-.search-info-containers {
+.event-info {
   background-color: rgb(100, 100, 100);
   border-radius: 25px;
   box-shadow: 0 0 4px black;
@@ -14,109 +14,5 @@
 }
 
 .ticket-scrollers-container {
-  /* width: 100%; */
-  /* max-width: 900px; */
-  /* min-width: 100%; */
   margin: 0 auto;
-}
-
-.ticket-list-container {
-  width: 1000px;
-  /* min-width: 100%; */
-  /* max-width: 100%; */
-  max-height: 275px;
-  margin: 10px;
-}
-
-.ticket-list-title {
-  background-color: rgb(100, 100, 100);
-  /* border-radius: 30% 30% 0 0; */
-}
-
-.ticket-scroller-container {
-  border: 1px solid rgb(100, 100, 100);
-  overflow-x: auto;
-  min-height: 100px;
-  position: relative;
-}
-
-.loader-container {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  width: 100%;
-  height: 100%;
-  position: absolute;
-}
-
-.custom-loader {
-  width: 50px; /* Adjust the size as needed */
-  height: 50px; /* Adjust the size as needed */
-  border: 4px solid rgba(0, 0, 0, 0.1); /* Adjust the border color */
-  border-top: 4px solid #3498db; /* Adjust the color of the loader */
-  border-radius: 50%;
-  animation: spin 1s linear infinite;
-}
-
-@keyframes spin {
-  0% { transform: rotate(0deg); }
-  100% { transform: rotate(360deg); }
-}
-
-.ticket-container {
-  display: flex;
-  height: 100%;
-  align-items: center;
-  padding: 20px 10px 20px 10px;
-}
-
-.ticket-tile {
-  height: 75%;
-  min-height: 200px;
-  max-height: 400px;
-  aspect-ratio: 1;
-  margin: 0 10px 0 10px;
-  font-size: x-small;
-  border-radius: 12px;
-  box-shadow: 0 0 4px black;
-  display: flex;
-  flex-direction: column;
-  text-decoration: none;
-}
-
-.gametime {
-  background-color: rgb(14, 14, 14);
-  color: white;
-}
-
-.tickpick {
-  background-color: white;
-  color: black;
-}
-
-.seat-view {
-  justify-content: center;
-  overflow: hidden;
-  border-radius: 12px 12px 0 0;
-  flex: 1;
-  display: flex;
-  align-items: center;
-}
-
-.seat-view img {
-  width: 175%;
-  height: auto;
-  object-fit: cover;
-}
-
-.ticket-info {
-  height: 25%;
-  display: grid;
-  grid-template-columns: repeat(3, 1fr);
-  align-items: center;
-}
-
-.ticket-app img {
-  width: 50%;
-  height: auto;
 }
