+++ frontend/src/components/main/TicketsPage/index.tsx
@@ -1,7 +1,10 @@
 import React, { useState, useEffect } from 'react';
-import TicketScroller from './TicketScroller';
+import TicketScroller from './ticketScroller';
 import './index.css';
 
+/**
+ * Represents the page to display the available tickets for an event.
+ */
 const TicketsPage = () => {
   const [valueTix, setValueTix] = useState([]);
   const [cheapestTix, setCheapestTix] = useState([]);
@@ -25,14 +28,14 @@ const TicketsPage = () => {
 
   return (
     <>
-      <div className='search-info'>
-        <div className='search-info-containers datetime'>
+      <div className='event-info-container'>
+        <div className='event-info datetime'>
           <span>Mon 10/28 7:30 PM</span>
         </div>
-        <div className='search-info-containers matchup'>
+        <div className='event-info matchup'>
           <span>Milwaukee Bucks @ Boston Celtics</span>
         </div>
-        <div className='search-info-containers ticket-quantity'>
+        <div className='event-info ticket-quantity'>
           <span>Quantity: 2</span>
         </div>
       </div>
