+++ frontend/src/components/main/TicketsPage/ticketScroller/index.css
@@ -1,0 +1,46 @@
+.ticket-list-container {
+  width: 1000px;
+  max-height: 275px;
+  margin: 10px;
+}
+
+.ticket-list-title {
+  background-color: rgb(100, 100, 100);
+}
+
+.ticket-scroller-container {
+  border: 1px solid rgb(100, 100, 100);
+  overflow-x: auto;
+  min-height: 100px;
+  position: relative;
+}
+
+.loader-container {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  width: 100%;
+  height: 100%;
+  position: absolute;
+}
+
+.custom-loader {
+  width: 50px;
+  height: 50px;
+  border: 4px solid rgba(0, 0, 0, 0.1);
+  border-top: 4px solid #3498db;
+  border-radius: 50%;
+  animation: spin 1s linear infinite;
+}
+
+@keyframes spin {
+  0% { transform: rotate(0deg); }
+  100% { transform: rotate(360deg); }
+}
+
+.tickets-container {
+  display: flex;
+  height: 100%;
+  align-items: center;
+  padding: 20px 10px 20px 10px;
+}
