+++ frontend/src/components/main/sidePanel/index.tsx
@@ -1,5 +1,8 @@
 import './index.css';
 
+/**
+ * Represents the side panel component.
+ */
 const SidePanel = () => <div className='side-panel'></div>;
 
 export default SidePanel;
