+++ frontend/src/index.tsx
@@ -1,7 +1,8 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
+import { BrowserRouter as Router } from 'react-router-dom';
 import { HelmetProvider } from 'react-helmet-async';
-import TicketTrack from './components/TicketTrack';
+import TicketTrack from './components/tickettrack';
 import './index.css';
 
 const rootElement = document.getElementById('root');
@@ -10,9 +11,11 @@ if (rootElement) {
   const root = ReactDOM.createRoot(rootElement);
   root.render(
     <React.StrictMode>
-      <HelmetProvider>
-        <TicketTrack />
-      </HelmetProvider>
+      <Router>
+        <HelmetProvider>
+          <TicketTrack />
+        </HelmetProvider>
+      </Router>
     </React.StrictMode>,
   );
 } else {
